Moves spooler tests to new o.e.s.t.e.queues.spooler package
Also adds suite class to this package to run all the tests.
